Avoid response decode-encode to avoid breaking large integers
The current implementation decodes a response, changes the id back to
the original request id, then reencodes the response.

This breaks for some JSON-RPC endpoints (Solana, for instance) that use
large integers (i.e. larger than will fit into a double), which get
reencoded as doubles.  For instance `"whatever":18446744073709551615`
will become `"whatever":1.18446744073709e+19`, and this both loses
precision and breaks parsers (such as some of Chainflip's Solana
deserializers) that properly expect to find a u64 (but *not* a float) as
that value.

This fix rewrites the lua handler to rewrite the *request* to include a
dummy `"id"`, then rewrites the dummy id in the response using string
substitution rather than a json decode-encode round trip.

It isn't perfect in that there's still a decode-encode pass in the
*request* (and so could break *requests* with large integers) but that
seems to be much less common than *responses* with large integers
(although primarily my use case here is as a Solana proxy for
Chainflip).