Add support for JSON-RPC batch requests
JSON-RPC v2 allows batch requests by sending a request consisting of an
array of multiple subrequests.

This fixes processing of such requests (which currently fail in the
generate_cache_key call because it expects a dict, not an array), and
caches the individual subrequests rather than the overall batch request.